Add Display and Displayable

+++ source/runtime/frame.hpp
@@ -28,7 +28,7 @@ public:
 
     void push_stack(Value value) { operand_stack_.push_back(value); }
     Value pop_stack();
-    Value peek_stack(int32_t index = 0) const { return operand_stack_[operand_stack_.size() - 1 - index]; }
+    Value peek_stack(uint16_t index = 0) const { return operand_stack_[operand_stack_.size() - 1 - index]; }
 
     Frame(const Frame&) = delete;
     Frame& operator=(const Frame&) = delete;
